Combine SSInstallerActionCell with HostInstallerActionCell
